requirements.txt
anyio==4.6.2.post1
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.4.0
filetype==1.2.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.0
idna==3.10
lxml==5.3.0
pyotp==2.9.0
python-dotenv==1.0.1
requests==2.32.3
setuptools==75.1.0
sniffio==1.3.1
socksio==1.0.0
soupsieve==2.6
twikit==2.2.0
urllib3==2.2.3
wheel==0.44.0
